Refactor js files into src.
