+++ Dockerfile
@@ -10,4 +10,4 @@ ADD . /opt/otp-data-builder/
 
 RUN yarn install --immutable
 
-CMD ( dockerd-entrypoint.sh --log-level=error > /dev/null 2>&1 & ) && unset DOCKER_HOST && echo "Sleeping for 5 minutes." && sleep 300 && node index.js
+CMD ( dockerd-entrypoint.sh --log-level=error > /dev/null 2>&1 & ) && unset DOCKER_HOST && echo "Sleeping for 5 minutes." && sleep 300 && node src/index.js
