+++ configs/finland/config.js
@@ -1,4 +1,4 @@
-const mapSrc = require('../../utils/configUtils.js');
+const mapSrc = require('../../src/utils/configUtils.js');
 
 module.exports = {
   id: 'finland',
