+++ src/index.js
@@ -2,7 +2,7 @@ const {
   postSlackMessage,
   getStartBuildMessage,
 } = require('./utils/builderUtils.js');
-const { update } = require('./task/Update');
+const { update } = require('./tasks/Update');
 const { SPLIT_BUILD_TYPE, timezone } = require('./config.js');
 const logger = require('./logger');
