+++ src/tasks/BlobValidation.js
